ProfileActivity  - 갤러리에서 고른 사진이 onActivityResult에서 받지 못하던 에러 수정
