+++ app/src/main/java/com/tistory/jeongs0222/kagongapplication/ui/profile/ProfileActivity.kt
@@ -1,6 +1,10 @@
 package com.tistory.jeongs0222.kagongapplication.ui.profile
 
+import android.app.Activity
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.util.Log
 import androidx.lifecycle.Observer
 import com.tistory.jeongs0222.kagongapplication.R
 import com.tistory.jeongs0222.kagongapplication.databinding.ActivityProfileBinding
@@ -20,9 +24,15 @@ class ProfileActivity: BaseActivity<ActivityProfileBinding>() {
     private val fragmentProvider = FragmentProviderImpl(supportFragmentManager) as FragmentProvider
     private val messageProvider = MessageProviderImpl(this@ProfileActivity) as MessageProvider
     private val dbHelperProvider = DBHelperProviderImpl(this@ProfileActivity) as DBHelperProvider
+    private val imageCropProvider = ImageCropProviderImpl(this@ProfileActivity) as ImageCropProvider
 
     private val profileDetailFragment = ProfileDetailFragment()
     private val profileModifyFragment = ProfileModifyFragment()
+
+    private val PICK_FROM_GALLERY = 111
+    private val CROP_FROM_CAMERA = 222
+
+    private lateinit var dataUri: Uri
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -57,4 +67,30 @@ class ProfileActivity: BaseActivity<ActivityProfileBinding>() {
         viewDataBinding.pViewModel = profileViewModel
         viewDataBinding.lifecycleOwner = this@ProfileActivity
     }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+
+        if(requestCode == PICK_FROM_GALLERY) {
+
+            if(resultCode == Activity.RESULT_OK) {
+
+                if(data != null) {
+                    val temp = data.data!!
+
+                    imageCropProvider.cropImage(temp)
+                }
+
+            }
+
+        } else if(requestCode == CROP_FROM_CAMERA) {
+            if(data != null) {
+                dataUri = data.data!!
+
+                val cropFile = imageCropProvider.galleryAddPic()
+                profileViewModel.uploadProfileImage(cropFile)
+            }
+        }
+    }
+
 }
